Create Player, Team, and League Routes
Create routes for user to be directed to when a statkeeper label is
selected on the main screen.
The routes are for player, team, and league statkeepers and the one the
user is taken to is dependent on the one the user taps on the home
screen.
The display for the routes are not yet implemented.