Create PlayerStatControls
Add a controller to choose a stat and increase or decrease the amount to
update it by. Pressing plus button increases the amount to add or
decreases the amount to subtract, while pressing the minus button
decreases the amount to add or increases the amount to subtract.
Clicking on a stat places that stat into the controller to update.
This feature is not yet fully functional, as it doesn't update the stat
shown yet.

Format double amounts such as AVG and OPS to show three spaces after
decimal and remove leading zero.