Implement editing of stats
Implement update button so when it is pressed in the PlayerStatControls,
the selected stat's amount (along with any other affected stats) is
updated.
Create static stats labels in the PlayerStatsClass and a list of
changeable stats, so only when one of those are clicked on can they be
updated in the controls.