Create Team class and StatKeeper and Team repositories
Create Team class, that holds a list of players and stats such as wins,
losses, runs scored, and runs allowed.
Create Team and StatKeeper repositories to store and retrieve data of
teams and SK's in sqlite db.
In main page, query and display all SK's
in the db, and expand CreateSKDialog to create SK's for teams and
leagues as well.