Begin implementation of game control buttons
Add playResult buttons that cover each potential play result for the
batter, where only one can be selected at a time.
When a playResult is selected, and only then, the "Submit Play" button
appears. When that button is then clicked, the players' stats are
updated in the database, the playResult is set back to null, and the
next batter is displayed.